Added shell script to run the scala programs.
