+++ json_parser/Scala/JsonParsing.scala
@@ -1,12 +1,12 @@
 import collection.immutable.IndexedSeq
-import _root_.com.google.gson.Gson
-import _root_.com.google.gson.JsonObject
-import _root_.com.google.gson.JsonParser
+import com.google.gson.Gson
+import com.google.gson.JsonObject
+import com.google.gson.JsonParser
 
 case class wrapperObject(val json_string: Array[MyJsonObject])
 case class MyJsonObject(val id:Int ,val price:Int)
 
-object Test2 {
+object JsonParsing {
   
     def maxVal():Integer = {
       
@@ -35,12 +35,7 @@ object Test2 {
        
       time{println(maxVal())}
       
-      val mb = 1024*1024
       val runtime = Runtime.getRuntime
-      println("\nMemory in MB")
-      println("** Used Memory:  " + (runtime.totalMemory - runtime.freeMemory))
-      println("** Free Memory:  " + runtime.freeMemory)
-      println("** Total Memory: " + runtime.totalMemory)
-      println("** Max Memory:   " + runtime.maxMemory)       
+      println("** Used Memory:  " + (runtime.totalMemory - runtime.freeMemory))    
     }
 }
